move the header file to a component directory
